김제 드론 및 분전반 데이터 scheduler

+++ danusys-commons-api/src/main/java/com/danusys/web/commons/api/model/Facility.java
@@ -29,7 +29,7 @@ public class Facility implements Serializable {
     private Long facilitySeq;
 
     @Column(nullable = false)
-    private int facilityKind;
+    private Long facilityKind;
 
     @Column(nullable = false)
     private int facilityStatus;
